Made minor quality improvements
