+++ src/Components/data-collection-page/AllianceSelection.tsx
